+++ x/pki/keeper/noc_certificates_by_subject_key_id_test.go
@@ -62,3 +62,24 @@ func TestNocCertificatesBySubjectKeyIdGetAll(t *testing.T) {
 		nullify.Fill(keeper.GetAllNocCertificatesBySubjectKeyID(ctx)),
 	)
 }
+
+func TestAddGetRemoveNocCertificatesBySubjectKeyID(t *testing.T) {
+	keeper, ctx := keepertest.PkiKeeper(t, nil)
+
+	keeper.AddNocCertificatesBySubjectKeyID(ctx, types.NocCertificates{
+		SubjectKeyId: "skid",
+		Certs: []*types.Certificate{
+			{Subject: "s1", SubjectKeyId: "skid"},
+		},
+	})
+
+	got, found := keeper.GetNocCertificatesBySubjectKeyID(ctx, "skid")
+	require.True(t, found)
+	require.Len(t, got.Certs, 1)
+	require.Equal(t, "s1", got.Certs[0].Subject)
+
+	keeper.RemoveNocCertificatesBySubjectAndSubjectKeyID(ctx, "s1", "skid")
+
+	_, found = keeper.GetNocCertificatesBySubjectKeyID(ctx, "skid")
+	require.False(t, found)
+}
